Power Bi Project - HR & Jobs Dashboard
About the project : Every company has a HR Department which deals with various recruitment and placement related work of the company. Here in this project we have a huge dataset in from which we are going to extract insights which can be useful for HR department to work on and to gain knowledge about the recruitment process of the market. We have extracted many different insights which we will observe further in the report.